pyproject.toml
[project]
name = "20251228-youtube-download-webui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "orjson>=3.11.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pytubefix>=10.3.6",
    "scrapetube>=2.6.0",
    "uvicorn>=0.40.0",
    "yt-dlp>=2025.12.8",
]
